Fix rubocop offences at accessor text field & text field spec
