+++ lib/screen-object/elements.rb
@@ -16,6 +16,6 @@ require 'screen-object/accessors/element'
 require 'screen-object/accessors/button'
 require 'screen-object/accessors/image'
 require 'screen-object/accessors/text'
-require 'screen-object/accessors/textfield'
+require 'screen-object/accessors/text_field'
 require 'screen-object/accessors/checkbox'
 require 'screen-object/accessors/table'
